fix small marks

+++ learner_site/course/models.py
@@ -1,11 +1,9 @@
 from django.db import models
 
-# Create your models here.
-
 
 class City(models.Model):
     name = models.TextField()
-    population = models.BigIntegerField()
+    population = models.PositiveIntegerField()
 
     def __str__(self):
         return self.name
